pyproject.toml
[tool.poetry]
name = "til"
version = "0.1.0"
description = ""
authors = ["Johnywhisky <92sursumcorda@gmail.com>"]

[tool.poetry.dependencies]
aiohttp = "^3.8.3"
alembic = "^1.8.1"
boto3 = "^1.26.1"
botocore = "^1.29.1"
celery = "^5.2.7"
Django = "^4.1.3"
django-cors-headers = "^3.13.0"
django-filter = "^22.1"
django-storages = "^1.13.1"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
djangorestframework-camel-case = "^1.3.0"
drf-extra-fields = "^3.4.1"
drf-yasg = "^1.21.4"
fastapi = "^0.85.2"
firebase-admin = "^6.0.1"
gunicorn = "^20.1.0"
Pillow = "^9.3.0"
psycopg2-binary = "^2.9.5"
python-dotenv = "^0.21.0"
python = "^3.11"
redis = "^4.3.4"
requests = "^2.28.1"
SQLAlchemy = "^1.4.42"
uvicorn = "^0.19.0"
pytest = "^7.2.0"
pydantic = "^1.10.2"
beautifulsoup4 = "^4.11.1"
aiofiles = "^22.1.0"
bcrypt = "^4.0.1"

[tool.poetry.dev-dependencies]
model-bakery = "^1.9.0"
openpyxl = "^3.0.10"
pylint = "^2.15.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
